+++ mantidimaging/core/filters/rebin/__init__.py
@@ -1,4 +1,2 @@
-NAME = 'Rebin'
-
-from .rebin import execute, modes, _cli_register  # noqa:F401
-from .rebin_gui import _gui_register  # noqa: F401  # noqa:F821
+from .rebin import RebinFilter, modes, _cli_register  # noqa:F401
+FILTER_CLASS = RebinFilter
