+++ mantidimaging/core/filters/rebin/test/rebin_test.py
@@ -4,10 +4,8 @@ import numpy as np
 import numpy.testing as npt
 
 import mantidimaging.test_helpers.unit_test_helper as th
-
+from mantidimaging.core.filters.rebin import RebinFilter
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-
-from mantidimaging.core.filters import rebin
 
 
 class RebinTest(unittest.TestCase):
@@ -26,7 +24,7 @@ class RebinTest(unittest.TestCase):
         val = None
         mode = 'nearest'
 
-        result = rebin.execute(images, val, mode)
+        result = RebinFilter()._filter_func(images, val, mode)
 
         npt.assert_equal(result, control)
         npt.assert_equal(images, control)
@@ -37,7 +35,7 @@ class RebinTest(unittest.TestCase):
         mode = 'nearest'
         val = -1
 
-        result = rebin.execute(images, val, mode)
+        result = RebinFilter()._filter_func(images, val, mode)
 
         npt.assert_equal(result, control)
         npt.assert_equal(images, control)
@@ -48,7 +46,7 @@ class RebinTest(unittest.TestCase):
         mode = 'nearest'
         val = 0
 
-        result = rebin.execute(images, val, mode)
+        result = RebinFilter()._filter_func(images, val, mode)
 
         npt.assert_equal(result, control)
         npt.assert_equal(images, control)
@@ -81,7 +79,7 @@ class RebinTest(unittest.TestCase):
         expected_x = int(images.shape[1] * val)
         expected_y = int(images.shape[2] * val)
 
-        result = rebin.execute(images, val, mode)
+        result = RebinFilter()._filter_func(images, val, mode)
 
         npt.assert_equal(result.shape[1], expected_x)
         npt.assert_equal(result.shape[2], expected_y)
@@ -124,7 +122,7 @@ class RebinTest(unittest.TestCase):
         expected_x = int(val[0])
         expected_y = int(val[1])
 
-        result = rebin.execute(images, rebin_param=val, mode=mode)
+        result = RebinFilter()._filter_func(images, rebin_param=val, mode=mode)
 
         npt.assert_equal(result.shape[1], expected_x)
         npt.assert_equal(result.shape[2], expected_y)
@@ -149,7 +147,7 @@ class RebinTest(unittest.TestCase):
 
         cached_memory = get_memory_usage_linux(kb=True)[0]
 
-        result = rebin.execute(images, val, mode)
+        result = RebinFilter()._filter_func(images, val, mode)
 
         self.assertLess(
             get_memory_usage_linux(kb=True)[0], cached_memory * 2)
